Fixed more windows warnings

+++ include/FixedSizeString.h
@@ -30,7 +30,7 @@ class   VirtualString  {
 
 public:
 
-    using size_type = unsigned int;
+    using size_type = size_t;
     using value_type = char;
     using pointer = value_type *;
     using const_pointer = const value_type *;
